Rename ConfigProvider -> PropProvider
Also remove old todo (.md) files.
