+++ src/components/PropProvider/README.md
@@ -1,4 +1,4 @@
-# ConfigProvider
+# PropProvider
 
 A Provider component provides child Blue components with config values.
