+++ src/components/PropProvider/docs/configConnect.md
@@ -1,12 +1,12 @@
 # configConnect
 
-Higher-order function that "connects" a React component with Blue's config, generated by the [ConfigProvider](./Provider.md). This pattern is similar to [Redux's connect](https://github.com/reduxjs/react-redux).
+Higher-order function that "connects" a React component with Blue's config, generated by the [PropProvider](./Provider.md). This pattern is similar to [Redux's connect](https://github.com/reduxjs/react-redux).
 
 ## Examples
 
 ```jsx
 import React, { Component } from 'react'
-import { configConnect } from '@helpscout/blue/ConfigProvider'
+import { configConnect } from '@helpscout/blue/PropProvider'
 
 const MyComponent extends Component {
   ...
