--- src/components/Tooltip/configs.js
@@ -1,3 +1,0 @@
-export default {
-  zIndex: 999,
-}
