+++ src/components/index.js
@@ -21,7 +21,7 @@ export { default as Choice } from './Choice'
 export { default as ChoiceGroup } from './ChoiceGroup'
 export { default as CloseButton } from './CloseButton'
 export { default as Code } from './Code'
-export { default as ConfigProvider } from './ConfigProvider'
+export { default as PropProvider } from './PropProvider'
 export { default as Collapsible } from './Collapsible'
 export { default as DetailList } from './DetailList'
 export { default as Dropdown } from './Dropdown'
