--- todos/README.md
@@ -1,9 +1,0 @@
-# Todos
-
-In this directory is a collection of todos (mostly design/dev notes).
-
-## Table of contents
-
-* [Components (React)](./components.md)
-* [Design](./design.md)
-* [Workflow](./workflow.md)
