--- todos/components.md
@@ -1,34 +1,0 @@
-# Components
-
-Below is a (working) list of components we need to build for Blue.
-
-Note: Beacon (v2) requires every one of these.
-
-* [x] Avatar
-* [x] AvatarStack
-* [x] Button `*`
-* [x] Badge
-* [x] Card
-* [x] CardBlock
-* [x] Checkbox/Radio
-* [x] Collapsible
-* [x] Drop
-* [x] Dropdown
-* [x] Heading
-* [x] Icon
-* [x] Image
-* [x] Input
-* [x] Label
-* [x] Link
-* [x] Loader (...)
-* [x] Modal `**`
-* [x] Overlay
-* [ ] Popover `*` (Need to leverage Drop)
-* [x] Portal `*`
-* [x] Select
-* [ ] Spinner
-* [x] Text
-* [ ] Tooltip `*` (Need to leverage Drop)
-* [ ] Video `***` (Difficult due to the variety of external video sources.)
-
-`*` = These will be hard(er) to code. The more stars, the harder it is.
