--- todos/design.md
@@ -1,58 +1,0 @@
-# Design
-
-Below is a (working) list of design notes/todos.
-
-**[Link to Sketch file on Dropbox](https://www.dropbox.com/s/ycy3ro44reajybr/Blue.sketch?dl=0)**
-
-
-## Input
-
-* [ ] Label styles
-* [ ] Sizes (name + style properties for each)
-* [ ] Error styles (Label, helper text, :hover, :active, :focus)
-* [ ] Success styles (Label, helper text, :hover, :active, :focus)
-* [ ] Warning styles (Label, helper text, :hover, :active, :focus)
-
-
-## Badge (Lozenge)
-
-* [ ] Sizes (name + style properties for each)
-* [ ] Error styles
-* [ ] Success styles
-* [ ] Warning styles
-
-
-## Button
-
-* [ ] Sizes (name + style properties for each)
-* [ ] Error styles (Label, helper text, :hover, :active, :focus)
-* [ ] Success styles (Label, helper text, :hover, :active, :focus)
-* [x] Warning styles (Label, helper text, :hover, :active, :focus)
-
-
-## DotIndicator (open to names, haha)
-
-* [x] States (name + style properties for each. E.g online, offline, idle?)
-
-
-## Headings
-
-* [ ] Sizes/styles (name + style properties for each)
-
-
-## Text
-
-* [ ] Sizes/styles (name + style properties for each)
-
-
-## Icons
-
-* [x] `.svg` Export of all icons
-
-
-## CloseButton
-
-(Used in Modal for the Close UI)
-
-* [ ] Sizes / alt styles?
-* [ ] Designs for `:focus`, `:hover`, `:active` (and any combo of those 3, if applicable)
